+++ packages/format/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @edge-runtime/format
 
+## 3.0.0
+
+### Major Changes
+
+- use MIT license ([#909](https://github.com/vercel/edge-runtime/pull/909))
+
 ## 2.2.1
 
 ### Patch Changes